Move DSU packet helpers to protocol package
